+++ CHANGELOG.md
@@ -10,6 +10,16 @@
 You should also include the user name that made the change.
 -->
 
+## 13.x.x (unreleased)
+
+### Improvements
+- クリップ、チャンネルページに共有ボタンを追加
+- ドライブの「URLからアップロード」で、content-dispositionのfilenameがあればそれをファイル名に
+- サーバーのパフォーマンスを改善
+
+### Bugfixes
+-
+
 ## 13.9.1 (2023/03/03)
 
 ### Bugfixes